requirements.txt
Django==4.0
python-dotenv==0.20.0
celery
redis
crispy-bootstrap5